+++ lib/src/core/transaction/secp256k1_sighash_all_builder.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:ckb_sdk_dart/ckb_crypto.dart';
 import 'package:ckb_sdk_dart/src/core/transaction/script_group.dart';
 import 'package:ckb_sdk_dart/src/core/type/transaction.dart';
-import 'package:ckb_sdk_dart/src/core/type/utils/serializer.dart';
+import 'package:ckb_sdk_dart/src/core/utils/serializer.dart';
 import 'package:ckb_sdk_dart/src/core/type/witness.dart';
 import 'package:ckb_sdk_dart/src/crypto/sign.dart';
 import 'package:ckb_sdk_dart/src/serialization/fixed/uint64.dart';
@@ -19,18 +19,14 @@ class Secp256k1SighashAllBuilder {
   void sign(ScriptGroup scriptGroup, String privateKey) {
     var groupWitnesses = [];
     if (_transaction.witnesses.length < _transaction.inputs.length) {
-      throw Exception(
-          'Transaction witnesses count must not be smaller than inputs count');
+      throw Exception('Transaction witnesses count must not be smaller than inputs count');
     }
     if (scriptGroup.inputIndexes.isEmpty) {
       throw Exception('Need at least one witness!');
     }
-    scriptGroup.inputIndexes
-        .forEach((index) => groupWitnesses.add(_transaction.witnesses[index]));
+    scriptGroup.inputIndexes.forEach((index) => groupWitnesses.add(_transaction.witnesses[index]));
 
-    for (var i = _transaction.inputs.length;
-        i < _transaction.witnesses.length;
-        i++) {
+    for (var i = _transaction.inputs.length; i < _transaction.witnesses.length; i++) {
       groupWitnesses.add(_transaction.witnesses[i]);
     }
     if (groupWitnesses[0] is! Witness) {
@@ -57,11 +53,9 @@ class Secp256k1SighashAllBuilder {
     var message = blake2b.doFinalString();
 
     Witness signedWitness = groupWitnesses[0];
-    signedWitness.lock = listToHex(
-        Sign.signMessage(hexToList(message), privateKey).getSignature());
+    signedWitness.lock = listToHex(Sign.signMessage(hexToList(message), privateKey).getSignature());
 
-    _transaction.witnesses[scriptGroup.inputIndexes[0]] =
-        listToHex(Serializer.serializeWitnessArgs(signedWitness).toBytes());
+    _transaction.witnesses[scriptGroup.inputIndexes[0]] = listToHex(Serializer.serializeWitnessArgs(signedWitness).toBytes());
   }
 
   Transaction buildTx() {