+++ lib/src/core/type/transaction.dart
@@ -9,7 +9,7 @@ import '../../../ckb_serialization.dart';
 import 'cell_dep.dart';
 import 'cell_input.dart';
 import 'cell_output.dart';
-import 'utils/serializer.dart';
+import '../utils/serializer.dart';
 
 class Transaction {
   String version;
@@ -21,38 +21,19 @@ class Transaction {
   List<String> outputsData;
   List<dynamic> witnesses;
 
-  Transaction(
-      {this.version,
-      this.hash,
-      this.cellDeps,
-      this.headerDeps,
-      this.inputs,
-      this.outputs,
-      this.outputsData,
-      this.witnesses});
+  Transaction({this.version, this.hash, this.cellDeps, this.headerDeps, this.inputs, this.outputs, this.outputsData, this.witnesses});
 
   factory Transaction.fromJson(Map<String, dynamic> json) {
     if (json == null) return null;
     return Transaction(
         version: json['version'],
         hash: json['hash'],
-        cellDeps: (json['cell_deps'] as List)
-            ?.map((cellDep) => CellDep.fromJson(cellDep))
-            ?.toList(),
-        headerDeps: (json['header_deps'] as List)
-            ?.map((headerDep) => headerDep?.toString())
-            ?.toList(),
-        inputs: (json['inputs'] as List)
-            ?.map((input) => CellInput.fromJson(input))
-            ?.toList(),
-        outputs: (json['outputs'] as List)
-            ?.map((output) => CellOutput.fromJson(output))
-            ?.toList(),
-        outputsData: (json['outputs_data'] as List)
-            ?.map((outputData) => outputData.toString())
-            ?.toList(),
-        witnesses:
-            (json['witnesses'] as List)?.map((witness) => witness)?.toList());
+        cellDeps: (json['cell_deps'] as List)?.map((cellDep) => CellDep.fromJson(cellDep))?.toList(),
+        headerDeps: (json['header_deps'] as List)?.map((headerDep) => headerDep?.toString())?.toList(),
+        inputs: (json['inputs'] as List)?.map((input) => CellInput.fromJson(input))?.toList(),
+        outputs: (json['outputs'] as List)?.map((output) => CellOutput.fromJson(output))?.toList(),
+        outputsData: (json['outputs_data'] as List)?.map((outputData) => outputData.toString())?.toList(),
+        witnesses: (json['witnesses'] as List)?.map((witness) => witness)?.toList());
   }
 
   Map<String, dynamic> toJson() {
@@ -104,8 +85,7 @@ class Transaction {
     for (var i = 1; i < witnesses.length; i++) {
       Uint8List bytes;
       if (witnesses[i] is Witness) {
-        bytes =
-            Serializer.serializeWitnessArgs(witnesses[i] as Witness).toBytes();
+        bytes = Serializer.serializeWitnessArgs(witnesses[i] as Witness).toBytes();
       } else {
         bytes = hexToList(witnesses[i] as String);
       }
@@ -113,27 +93,17 @@ class Transaction {
       blake2b.update(bytes);
     }
     var message = blake2b.doFinalString();
-    (witnesses[0] as Witness).lock = listToHex(
-        Sign.signMessage(hexToList(message), privateKey).getSignature());
+    (witnesses[0] as Witness).lock = listToHex(Sign.signMessage(hexToList(message), privateKey).getSignature());
 
     var signedWitness = [];
     for (Object witness in witnesses) {
       if (witness is Witness) {
-        signedWitness
-            .add(listToHex(Serializer.serializeWitnessArgs(witness).toBytes()));
+        signedWitness.add(listToHex(Serializer.serializeWitnessArgs(witness).toBytes()));
       } else {
         signedWitness.add(witness);
       }
     }
 
-    return Transaction(
-        version: version,
-        hash: txHash,
-        cellDeps: cellDeps,
-        headerDeps: headerDeps,
-        inputs: inputs,
-        outputs: outputs,
-        outputsData: outputsData,
-        witnesses: signedWitness);
+    return Transaction(version: version, hash: txHash, cellDeps: cellDeps, headerDeps: headerDeps, inputs: inputs, outputs: outputs, outputsData: outputsData, witnesses: signedWitness);
   }
 }